backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
httpx>=0.27.0
anthropic>=0.25.0
openai>=1.12.0
websockets>=12.0
